Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref>=3.8.1",
    "cloudinary>=1.44.1",
    "django>=5.2.3",
    "django-cloudinary-storage>=0.3.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "dj-database-url>=2.1.0",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "pillow>=11.2.1",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.1.1",
]
